image_processing/requirements.txt
Click==7.0
dlib==19.16.0
face-recognition==1.2.3
face-recognition-models==0.3.0
imutils==0.5.1
numpy==1.15.3
opencv-contrib-python==3.4.3.18
Pillow==5.3.0
pytesseract==0.2.5
keras==2.0.5
tensorflow==1.11
pandas==0.19.1
h5py==2.7.0
statistics
cycler==0.10.0
kiwisolver==1.0.1
matplotlib==3.0.1
pyparsing==2.2.2
python-dateutil==2.7.5
six==1.11.0
